Hotfix: Recurrence definitions weren't autoloaded via psr-4
Multi-class file splitted.
