+++ src/Definition/Response/PreAuthState.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace GoPay\Definition\Response;
+
+class PreAuthState
+{
+
+    const REQUESTED = 'REQUESTED';
+    const AUTHORIZED = 'AUTHORIZED';
+    const CAPTURED = 'CAPTURED';
+    const CANCELED = 'CANCELED';
+}
