+++ src/Definition/Response/RecurrenceState.php
@@ -9,13 +9,3 @@ class RecurrenceState
     const STARTED = 'STARTED';
     const STOPPED = 'STOPPED';
 }
-
-class PreAuthState
-{
-
-    const REQUESTED = 'REQUESTED';
-    const AUTHORIZED = 'AUTHORIZED';
-    const CAPTURED = 'CAPTURED';
-    const CANCELED = 'CANCELED';
-}
-
